feat(backend): add execution and key routes, health, and the demo seed
The seed creates one flagged demo account and then runs its workflows through
the real engine against a local sink, so every execution in the dashboard is a
genuine run — including one that exhausts all three attempts. Durations are
stored in microseconds because a mocked action finishes in well under a
millisecond.

+++ backend/app/schemas/api_key.py
@@ -44,5 +44,5 @@ class StatsOut(BaseModel):
     succeeded: int
     failed: int
     success_rate: float | None
-    avg_duration_ms: int | None
+    avg_duration_ms: float | None
     series: list[dict]
